Refactoring

+++ java-basics/src/main/java/com/gb/Lesson012/Calculator/View.java
@@ -2,7 +2,7 @@ package com.gb.Lesson012.Calculator;
 
 public interface View {
 
-    int getValue(String message);
+    float getValue(String message);
 
     void print(String message);
