+++ java-basics/src/main/java/com/gb/Lesson012/Calculator/ViewImplementation.java
@@ -15,9 +15,9 @@ public class ViewImplementation implements View {
     }
 
     @Override
-    public int getValue(String message) {
+    public float getValue(String message) {
         out.print(message);
-        return scn.nextInt();
+        return scn.nextFloat();
     }
 
     @Override
